LangevinIntegrators

Documentation for LangevinIntegrators.

Initial conditions

LangevinIntegrators.initialize_initcond — Function

Function to initialize the init_cond Note si il n'y as pas ce qu'il faut ça va échouer silenciement, il faut mettre un verbose pour montrer ce qui est utilisé

Run the trajectories

LangevinIntegrators.TrajsParams — Type

TrajsParams(;n_steps = 10^4, n_save_iters=1)

Set options for samplers.

Fields

n_steps - Set the number of iterations of the integrator

n_trajs - Set the number of trajectories to run

nsaveiters - Set the frequency at which iterations are saved. If nsaveiters=1, every iteration is saved. If nsaveiters=n_steps, only the final iteration is saved.

savefilenamepattern - The pattern for saving the trajectories. If this include a *, that will be replaced by the id of the traj.

verbose - Verbose argument

LangevinIntegrators.run_trajectory! — Function

run_trajectory!(state, integrator)

evolve in time one trajectory starting at state and integrated by integrator

Fields

state - Initial State

integrator - Friction matrix

LangevinIntegrators.run_trajectories — Function

For launching a bunch of trajectories. For distributed computing see examples

Index

LangevinIntegrators.ABOBA

LangevinIntegrators.ABOBA_Hidden

LangevinIntegrators.BAOAB

LangevinIntegrators.BBK

LangevinIntegrators.BBK_Kernel

LangevinIntegrators.EM

LangevinIntegrators.EM_Hidden

LangevinIntegrators.EM_Kernel

LangevinIntegrators.ForceFromBasis

LangevinIntegrators.ForceFromPotential

LangevinIntegrators.ForceFromScipySplines

LangevinIntegrators.ForceFromSplines

LangevinIntegrators.GJ

LangevinIntegrators.GJF_Kernel

LangevinIntegrators.OBABO

LangevinIntegrators.PositionVerlet

LangevinIntegrators.SeparateSpace

LangevinIntegrators.TrajsParams

LangevinIntegrators.VEC

LangevinIntegrators.VelocityVerlet

LangevinIntegrators.GJF

LangevinIntegrators.LWall

LangevinIntegrators.Quadratic

LangevinIntegrators.UWall

LangevinIntegrators.initialize_initcond

LangevinIntegrators.run_trajectories

LangevinIntegrators.run_trajectory!