Integrators

Hamiltonian dynamic

LangevinIntegrators.VelocityVerlet — Type

VelocityVerlet(force, M, Δt)

Set up the velocity Verlet integrator.

Fields

force - In place gradient of the potential

M - Mass (either scalar or vector)

Δt - Time step

LangevinIntegrators.PositionVerlet — Type

PositionVerlet(force, M, Δt)

Set up the position Verlet integrator.

Fields

force - In place gradient of the potential

M - Mass (either scalar or vector)

Δt - Time step

Overdamped dynamics

LangevinIntegrators.EM — Type

EM(force, β, Δt)

Set up the EM integrator for overdamped Langevin.

Fields

force - In place gradient of the potential

β - Inverse temperature

Δt - Time step

Langevin dynamics

LangevinIntegrators.BBK — Type

BBK(force, β, γ, M, Δt)

Set up the BBK integrator for inertial Langevin.

Fields

force - In place gradient of the potential

β - Inverse temperature

γ - Damping Coefficient

M - Mass (either scalar or vector)

Δt - Time step

LangevinIntegrators.VEC — Type

VEC(force, β, γ, M, Δt)

Set up the Vanden-Eijnden Ciccotti integrator for inertial Langevin. Taken from "Second-order integrators for Langevin equations with holonomic constraints" doi: 10.1016/j.cplett.2006.07.086

Fields

force - In place gradient of the potential

β - Inverse temperature

γ - Damping Coefficient

M - Mass (either scalar or vector)

Δt - Time step

LangevinIntegrators.GJ — Type

GJ(force, β, γ, M, Δt, type)

Set up the various GJ integrator for inertial Langevin. Use type to select the approriate integrator

Fields

force - In place gradient of the potential

β - Inverse temperature

γ - Damping Coefficient

M - Mass (either scalar or vector)

Δt - Time step

type - Choice of the integrator should be one of "I","II","III","IV","V","VI"

LangevinIntegrators.GJF — Function

GJF(force, β, γ, M, Δt)

Set up the G-JF integrator for inertial Langevin.

Fields

force - In place gradient of the potential

β - Inverse temperature

γ - Damping Coefficient

M - Mass (either scalar or vector)

Δt - Time step

LangevinIntegrators.BAOAB — Type

BAOAB(force, β, γ, M, Δt)

Set up the BAOAB integrator for inertial Langevin.

Fields

force - In place gradient of the potential

β - Inverse temperature

γ - Damping Coefficient

M - Mass (either scalar or vector)

Δt - Time step

LangevinIntegrators.OBABO — Type

OBABO(force, β, γ, M, Δt)

Set up the OBABO integrator for inertial Langevin.

Fields

force - In place gradient of the potential

β - Inverse temperature

γ - Damping Coefficient

M - Mass (either scalar or vector)

Δt - Time step

LangevinIntegrators.ABOBA — Type

ABOBA(force, β, γ, M, Δt)

Set up the ABOBA integrator for inertial Langevin.

Fields

force - In place gradient of the potential

β - Inverse temperature

γ - Damping Coefficient

M - Mass (either scalar or vector)

Δt - Time step

Generalized Langevin dynamics

Missing docstring.

Missing docstring for EM_Kernel. Check Documenter's build log for details.

LangevinIntegrators.BBK_Kernel — Type

BBK_Kernel(force, β, γ, M, Δt)

Set up the BBK_Kernel integrator for generalized Langevin.

Adapted from Iterative Reconstruction of Memory Kernels Gerhard Jung,,†,‡ Martin Hanke,, and Friederike Schmid*,†

Fields

force - In place gradient of the potential

β - Inverse temperature

γ - Damping Coefficient

M - Mass (either scalar or vector)

Δt - Time step

LangevinIntegrators.GJF_Kernel — Type

GJF_Kernel(force, β, γ, M, Δt)

Set up the G-JF integrator for inertial Langevin.

Adapted from Iterative Reconstruction of Memory Kernels Gerhard Jung,,†,‡ Martin Hanke,, and Friederike Schmid*,†

Fields

force - In place gradient of the potential

β - Inverse temperature

γ - Damping Coefficient

M - Mass (either scalar or vector)

Δt - Time step

Generalize Langevin dynamic with hidden variables

LangevinIntegrators.EM_Hidden — Type

EM_Hidden(force, β, Δt)

Set up the EM_Hidden integrator for underdamped Langevin with hidden variables.

Fields

force - In place gradient of the potential

A - Friction matrix

C - diffusion matrix

Δt - Time step

LangevinIntegrators.ABOBA_Hidden — Type

ABOBA_Hidden(force, β, γ, M, Δt)

Set up the ABOBA_Hidden integrator for underdamped Langevin with hidden variables.

Fields

force - In place gradient of the potential

β - Inverse temperature

γ - Damping Coefficient

M - Mass (either scalar or vector)

Δt - Time step