LangevinIntegrators

Documentation for LangevinIntegrators.

LangevinIntegrators.ABOBA

LangevinIntegrators.ABOBA_Hidden

LangevinIntegrators.BAOAB

LangevinIntegrators.BBK

LangevinIntegrators.BBK_Kernel

LangevinIntegrators.EM

LangevinIntegrators.ForceFromScipySplines

LangevinIntegrators.GJF_Kernel

LangevinIntegrators.SeparateSpace

LangevinIntegrators.TrajsParams

LangevinIntegrators.VEC

LangevinIntegrators.VelocityVerlet

LangevinIntegrators.initialize_initcond

LangevinIntegrators.run_trajectories

LangevinIntegrators.run_trajectory!

LangevinIntegrators.ABOBA — Method

ABOBA(force, β, γ, M, Δt)

Set up the ABOBA integrator for inertial Langevin.

Fields

force - In place gradient of the potential

β - Inverse temperature

γ - Damping Coefficient

M - Mass (either scalar or vector)

Δt - Time step

LangevinIntegrators.ABOBA_Hidden — Method

ABOBA_Hidden(force, β, γ, M, Δt)

Set up the ABOBA_Hidden integrator for underdamped Langevin with hidden variables.

Fields

force - In place gradient of the potential

β - Inverse temperature

γ - Damping Coefficient

M - Mass (either scalar or vector)

Δt - Time step

LangevinIntegrators.BAOAB — Method

BAOAB(force, β, γ, M, Δt)

Set up the BAOAB integrator for inertial Langevin.

Fields

force - In place gradient of the potential

β - Inverse temperature

γ - Damping Coefficient

M - Mass (either scalar or vector)

Δt - Time step

LangevinIntegrators.BBK — Method

BBK(force, β, γ, M, Δt)

Set up the BBK integrator for inertial Langevin.

Fields

force - In place gradient of the potential

β - Inverse temperature

γ - Damping Coefficient

M - Mass (either scalar or vector)

Δt - Time step

LangevinIntegrators.BBK_Kernel — Method

BBK_Kernel(force, β, γ, M, Δt)

Set up the BBK_Kernel integrator for generalized Langevin.

Adapted from Iterative Reconstruction of Memory Kernels Gerhard Jung,,†,‡ Martin Hanke,, and Friederike Schmid*,†

Fields

force - In place gradient of the potential

β - Inverse temperature

γ - Damping Coefficient

M - Mass (either scalar or vector)

Δt - Time step

LangevinIntegrators.EM — Method

EM(force, β, Δt)

Set up the EM integrator for overdamped Langevin.

Fields

force - In place gradient of the potential

β - Inverse temperature

Δt - Time step

LangevinIntegrators.ForceFromScipySplines — Type

Careful ForceFromScipySplines is not threadsafe

LangevinIntegrators.GJF_Kernel — Method

GJF_Kernel(force, β, γ, M, Δt)

Set up the G-JF integrator for inertial Langevin.

Adapted from Iterative Reconstruction of Memory Kernels Gerhard Jung,,†,‡ Martin Hanke,, and Friederike Schmid*,†

Fields

force - In place gradient of the potential

β - Inverse temperature

γ - Damping Coefficient

M - Mass (either scalar or vector)

Δt - Time step

LangevinIntegrators.SeparateSpace — Type

When we want to appy boundary conditions, define a SeparateSpace, that hold one BC or constraints per dimension. That can hold a constraints on some dimension if needed

LangevinIntegrators.TrajsParams — Method

TrajsParams(;n_steps = 10^4, n_save_iters=1)

Set options for samplers.

Fields

n_steps - Set the number of iterations of the integrator

n_trajs - Set the number of trajectories to run

nsaveiters - Set the frequency at which iterations are saved. If nsaveiters=1, every iteration is saved. If nsaveiters=n_steps, only the final iteration is saved.

savefilenamepattern - The pattern for saving the trajectories. If this include a *, that will be replaced by the id of the traj.

verbose - Verbose argument

LangevinIntegrators.VEC — Method

VEC(force, β, γ, M, Δt)

Set up the Vanden-Eijnden Ciccotti integrator for inertial Langevin. Taken from "Second-order integrators for Langevin equations with holonomic constraints" doi: 10.1016/j.cplett.2006.07.086

Fields

force - In place gradient of the potential

β - Inverse temperature

γ - Damping Coefficient

M - Mass (either scalar or vector)

Δt - Time step

LangevinIntegrators.VelocityVerlet — Type

Verlet(force, M, Δt)

Set up the Verlet integrator.

Fields

force - In place gradient of the potential

M - Mass (either scalar or vector)

Δt - Time step

LangevinIntegrators.initialize_initcond — Method

Function to initialize the init_cond Note si il n'y as pas ce qu'il faut ça va échouer silenciement, il faut mettre un verbose pour montrer ce qui est utilisé

LangevinIntegrators.run_trajectories — Method

For launching a bunch of trajectories. For distributed computing see examples

LangevinIntegrators.run_trajectory! — Method

run_trajectory!(state, integrator)

evolve in time one trajectory starting at state and integrated by integrator

Fields

state - Initial State

integrator - Friction matrix